pyproject.toml
[project]
name = "rest-trainings-api"
version = "0.1.0"
description = ""
authors = [
    {name = "AlexandrSmolyachkovGH",email = "alexandr.smolyachkov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "yoyo-migrations (>=9.0.0,<10.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiogram (>=3.19.0,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery[redis] (>=5.4.0,<6.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-order (>=1.3.0,<2.0.0)",
    "pytest-ordering (>=0.6,<0.7)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false
